+++ Nymph/src/main/kotlin/com/ambrosia/nymph/services/TableService.kt
@@ -28,20 +28,22 @@ class TableService(
 
     @Transactional
     fun editTable(businessId: Long, tableId: Long, tableDto: TableDto): TableDto {
-        businessRepository.findById(businessId)
-            .orElseThrow { EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId)) }
+        if (!businessRepository.existsById(businessId)) {
+            throw EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId))
+        }
         val table = tableRepository.findById(tableId)
             .orElseThrow { EntityNotFoundException(Table::class.java, mutableMapOf("id" to tableId)) }
         tableDto.number?.let { table.number = it }
-        tableDto.isAvailable?.let { table.isAvailable = it }
+        tableDto.available?.let { table.available = it }
         tableRepository.save(table)
         return table.toDto()
     }
 
     @Transactional
     fun deleteTable(businessId: Long, tableId: Long) {
-        businessRepository.findById(businessId)
-            .orElseThrow { EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId)) }
+        if (!businessRepository.existsById(businessId)) {
+            throw EntityNotFoundException(Business::class.java, mutableMapOf("id" to businessId))
+        }
         val table = tableRepository.findById(tableId)
             .orElseThrow { EntityNotFoundException(Table::class.java, mutableMapOf("id" to tableId)) }
         tableRepository.delete(table)
